refactor: 영화 예약 데이터를 위한 ReservationRepository 추가

+++ app/src/main/java/woowacourse/movie/model/pricing/PricingSystem.kt
@@ -8,5 +8,5 @@ interface PricingSystem : Serializable {
     fun calculatePrice(
         screening: Screening,
         quantity: Quantity,
-    ): Int
+    ): Long
 }
